fix Interfaces and abstract methods
