+++ library/Zend/Db/Sql/Delete.php
@@ -18,7 +18,7 @@ use Zend\Db\Adapter\StatementContainerInterface;
  *
  * @property Where $where
  */
-class Delete extends AbstractSql implements SqlInterface, PreparableSqlInterface
+class Delete extends AbstractPreparableSql
 {
     /**@#+
      * @const
