+++ library/Zend/Db/Sql/Update.php
@@ -18,7 +18,7 @@ use Zend\Db\Adapter\StatementContainerInterface;
  *
  * @property Where $where
  */
-class Update extends AbstractSql implements SqlInterface, PreparableSqlInterface
+class Update extends AbstractPreparableSql
 {
     /**@#++
      * @const
